multirom: Enhance stock compatibility (prevent mounting apps_log)
In N roms that use the Sony AOSP init implementation, the apps_log
partition is used to store some temporal data for the AOSP
recovery. This partition ist not used in AOSP roms, but by stock
roms, and modifications to it prevent booting stock roms if the
apps_log partition is not wiped before.

With this commit the fstab is modified when multirom is started
such that any line containing the apps_log string is removed prior
to booting the rom. Hence, modifications to the apps_log partition
are prevented for all roms (primary & secondary), which allows to
conveniently use stock and newer custom roms in parallel.